split main into cli and server packages
