+++ api/server.go
@@ -1,4 +1,4 @@
-package server
+package api
 
 import (
 	"bytes"
@@ -30,9 +30,6 @@ type EvalResponse struct {
 
 func Serve() {
 	log.Printf("Running server on port :%d\n", 5150)
-	// This is a sanity check since I'm not really very
-	// familiar with Go's http api yet
-	http.HandleFunc("/", handleRoot)
 	http.HandleFunc("/eval", handleEvaluate)
 	err := http.ListenAndServe(":5150", nil)
 	if errors.Is(err, http.ErrServerClosed) {
@@ -43,11 +40,6 @@ func Serve() {
 		log.Printf("Error: %s\n", err)
 		os.Exit(1)
 	}
-}
-
-func handleRoot(w http.ResponseWriter, r *http.Request) {
-	log.Println("This is the root route!")
-	io.WriteString(w, "Test")
 }
 
 func handleEvaluate(w http.ResponseWriter, r *http.Request) {
