+++ cli/main.go
@@ -1,0 +1,21 @@
+package main
+
+import (
+	"fmt"
+	"monkey-pl/repl"
+	"os"
+	"os/user"
+)
+
+func main() {
+	user, err := user.Current()
+
+	if err != nil {
+		fmt.Println("🙊 Oh no! There was an error! See below: 🙊")
+		panic(err)
+	}
+
+	fmt.Printf("🐵 Hello %s! Welcome to the Monkey programming language 🐵\n", user.Username)
+	fmt.Printf("🐵🍌 Try out some commands! Use :exit to exit the repl 🍌🐵\n\n")
+	repl.Start(os.Stdin, os.Stdout)
+}
