--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "monkey-pl/server"
-
-func main() {
-	server.Serve()
-}
