+++ server/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "monkey-pl/api"
+
+func main() {
+	api.Serve()
+}
